// frameworks/bridge/arkts_frontend/koala_mirror/arkoala-arkts/arkui/src/generated/arkui-graphics.ts
export class LengthMetricsInternal {
    public static fromPtr(ptr: KPointer): LengthMetrics {
        const obj : LengthMetrics = new LengthMetrics()
        obj.peer = new Finalizable(ptr, LengthMetrics.getFinalizer())
        return obj
    }
}
export class LengthMetrics implements MaterializedBase {
    peer?: Finalizable | undefined = undefined
    public getPeer(): Finalizable | undefined {
        return this.peer
    }
    get unit(): LengthUnit {
        return this.getUnit()
    }
    set unit(unit: LengthUnit) {
        this.setUnit(unit)
    }
    get value(): number {
        return this.getValue()
    }
    set value(value: number) {
        this.setValue(value)
    }
    static ctor_lengthmetrics(): KPointer {
        const retval  = ArkUIGeneratedNativeModule._LengthMetrics_ctor()
        return retval
    }
    constructor() {
        // Constructor does not have parameters.
        // It means that the static method call invokes ctor method as well
        // when all arguments are undefined.
        const ctorPtr : KPointer = LengthMetrics.ctor_lengthmetrics()
        this.peer = new Finalizable(ctorPtr, LengthMetrics.getFinalizer())
    }
    static getFinalizer(): KPointer {
        return ArkUIGeneratedNativeModule._LengthMetrics_getFinalizer()
    }
    public static px(value: number): LengthMetrics {
        const value_casted = value as (number)
        return LengthMetrics.px_serialize(value_casted)
    }
    public static vp(value: number): LengthMetrics {
        const value_casted = value as (number)
        return LengthMetrics.vp_serialize(value_casted)
    }
    public static fp(value: number): LengthMetrics {
        const value_casted = value as (number)
        return LengthMetrics.fp_serialize(value_casted)
    }
    public static percent(value: number): LengthMetrics {
        const value_casted = value as (number)
        return LengthMetrics.percent_serialize(value_casted)
    }
    public static lpx(value: number): LengthMetrics {
        const value_casted = value as (number)
        return LengthMetrics.lpx_serialize(value_casted)
    }
    public static resource(value: Resource): LengthMetrics {
        const value_casted = value as (Resource)
        return LengthMetrics.resource_serialize(value_casted)
    }
    private getUnit(): LengthUnit {
        return this.getUnit_serialize()
    }
    private setUnit(unit: LengthUnit): void {
        const unit_casted = unit as (LengthUnit)
        this.setUnit_serialize(unit_casted)
        return
    }
    private getValue(): number {
        return this.getValue_serialize()
    }
    private setValue(value: number): void {
        const value_casted = value as (number)
        this.setValue_serialize(value_casted)
        return
    }
    private static px_serialize(value: number): LengthMetrics {
        const retval  = ArkUIGeneratedNativeModule._LengthMetrics_px(value)
        const obj : LengthMetrics = LengthMetricsInternal.fromPtr(retval)
        return obj
    }
    private static vp_serialize(value: number): LengthMetrics {
        const retval  = ArkUIGeneratedNativeModule._LengthMetrics_vp(value)
        const obj : LengthMetrics = LengthMetricsInternal.fromPtr(retval)
        return obj
    }
    private static fp_serialize(value: number): LengthMetrics {
        const retval  = ArkUIGeneratedNativeModule._LengthMetrics_fp(value)
        const obj : LengthMetrics = LengthMetricsInternal.fromPtr(retval)
        return obj
    }
    private static percent_serialize(value: number): LengthMetrics {
        const retval  = ArkUIGeneratedNativeModule._LengthMetrics_percent(value)
        const obj : LengthMetrics = LengthMetricsInternal.fromPtr(retval)
        return obj
    }
    private static lpx_serialize(value: number): LengthMetrics {
        const retval  = ArkUIGeneratedNativeModule._LengthMetrics_lpx(value)
        const obj : LengthMetrics = LengthMetricsInternal.fromPtr(retval)
        return obj
    }
    private static resource_serialize(value: Resource): LengthMetrics {
        const thisSerializer : Serializer = Serializer.hold()
        thisSerializer.writeResource(value)
        const retval  = ArkUIGeneratedNativeModule._LengthMetrics_resource(thisSerializer.asBuffer(), thisSerializer.length())
        thisSerializer.release()
        const obj : LengthMetrics = LengthMetricsInternal.fromPtr(retval)
        return obj
    }
    private getUnit_serialize(): LengthUnit {
        const retval  = ArkUIGeneratedNativeModule._LengthMetrics_getUnit(this.peer!.ptr)
        return TypeChecker.LengthUnit_FromNumeric(retval)
    }
    private setUnit_serialize(unit: LengthUnit): void {
        ArkUIGeneratedNativeModule._LengthMetrics_setUnit(this.peer!.ptr, TypeChecker.LengthUnit_ToNumeric(unit))
    }
    private getValue_serialize(): number {
        const retval  = ArkUIGeneratedNativeModule._LengthMetrics_getValue(this.peer!.ptr)
        return retval
    }
    private setValue_serialize(value: number): void {
        ArkUIGeneratedNativeModule._LengthMetrics_setValue(this.peer!.ptr, value)
    }
}
// ...
export enum LengthUnit {
    PX = 0,
    VP = 1,
    FP = 2,
    PERCENT = 3,
    LPX = 4
}